requirements.txt
python-daemon == 2.1.1
socketIO-client == 0.7.0
